+++ scripts/macros.js
@@ -265,9 +265,9 @@ export {celestialResistance} from './macros/2024/speciesFeatures/aasimar/celesti
 // Cartographer
 export {adventurersAtlas} from './macros/2024/classFeatures/artificer/cartographer/adventurersAtlas.js';
 export {guidedPrecision} from './macros/2024/classFeatures/artificer/cartographer/guidedPrecision.js';
-//export {ingeniousMovement} from './macros/2024/classFeatures/artificer/cartographer/ingeniousMovement.js';
+export {ingeniousMovement} from './macros/2024/classFeatures/artificer/cartographer/ingeniousMovement.js';
 export {mappingMagic} from './macros/2024/classFeatures/artificer/cartographer/mappingMagic.js';
-//export {superiorAtlas, superiorAtlasEffect} from './macros/2024/classFeatures/artificer/cartographer/superiorAtlas.js';
+export {superiorAtlas, superiorAtlasEffect} from './macros/2024/classFeatures/artificer/cartographer/superiorAtlas.js';
 // Barbarian
 export {rage, rageRaging, rageUpkeep} from './macros/2024/classFeatures/barbarian/rage.js';
 export {unarmoredDefenseBarbarian} from './macros/2024/classFeatures/barbarian/unarmoredDefense.js';