+++ scripts/macros/2024/classFeatures/artificer/cartographer/adventurersAtlas.js
@@ -1,7 +1,7 @@
-import {actorUtils, dialogUtils, effectUtils, genericUtils, itemUtils, tokenUtils, workflowUtils} from '../../../../../utils.js';
+import {actorUtils, dialogUtils, effectUtils, genericUtils, itemUtils, socketUtils, tokenUtils, workflowUtils} from '../../../../../utils.js';
 async function makeMaps({trigger: {entity: item}, workflow}) {
     if (itemUtils.getConfig(item, 'requireTools') && !workflow.actor.items.some(i => i.system.type?.baseItem === 'cartographer')) {
-        genericUtils.notify('CHRISPREMADES.Macros.AdventurersAtlas.NeedTools', 'warn', {localize: true});
+        genericUtils.notify(genericUtils.format('CHRISPREMADES.Macros.AdventurersAtlas.NeedTools', {itemName: item.name}), 'warn');
         return;
     }
     let effectData = {
@@ -10,8 +10,7 @@ async function makeMaps({trigger: {entity: item}, workflow}) {
         origin: item.uuid,
         flags: {
             dae: {
-                stackable: 'noneName',
-                showIcon: true
+                stackable: 'noneName'
             }
         }
     };
@@ -23,7 +22,8 @@ async function makeMaps({trigger: {entity: item}, workflow}) {
         mode: 2,
         priority: 20
     }];
-    //if (itemUtils.getItemByIdentifier(workflow.actor, 'superiorAtlas')) effectUtils.addMacro(effectData, 'midi.actor', ['superiorAtlasEffect']);
+    genericUtils.setProperty(effectData, 'flags.dae.showIcon', true);
+    if (itemUtils.getItemByIdentifier(workflow.actor, 'superiorAtlas')) effectUtils.addMacro(effectData, 'midi.actor', ['superiorAtlasEffect']);
     await Promise.allSettled(workflow.targets.map(t => effectUtils.createEffect(t.actor, effectData, {
         identifier: 'adventurersAtlas',
         parentEntity: source,
@@ -32,28 +32,25 @@ async function makeMaps({trigger: {entity: item}, workflow}) {
 }
 async function rest({trigger: {entity: item}, actor}) {
     if (itemUtils.getConfig(item, 'requireTools') && !actor.items.some(i => i.system.type?.baseItem === 'cartographer')) {
-        genericUtils.notify('CHRISPREMADES.Macros.AdventurersAtlas.NeedTools', 'warn', {localize: true});
+        genericUtils.notify(genericUtils.format('CHRISPREMADES.Macros.AdventurersAtlas.NeedTools', {itemName: item.name}), 'warn');
         return;
     }
     let token = actorUtils.getFirstToken(actor);
     if (!token) return;
-    let near = tokenUtils.findNearby(token, 5, 'ally', {includeToken: true});
-    if (!near?.length) return;
-    let targets;
-    if (near.length === 1) {
-        targets = [near[0], token];
-        if (dialogUtils.confirm(item.name, genericUtils.format('CHRISPREMADES.Dialog.UseOn', {itemName: item.name, tokenName: near[0].name})))
-            await workflowUtils.syntheticItemRoll(item, targets);
-    } else {
-        targets = await dialogUtils.selectTargetDialog(item.name, genericUtils.format('CHRISPREMADES.Dialog.Use', {itemName: item.name}), near, {
-            type: 'multiple',
-            maxAmount: Math.max(1 + actor.system.abilities.int.mod, 2),
-            minAmount: 2
-        });
-        if (!targets || !targets[0]?.length) return;
-        targets = targets[0];
-        await workflowUtils.syntheticItemRoll(item, targets);
-    }
+    let range = itemUtils.getConfig(item, 'enforceRange') ? 5 : 60;
+    let near = tokenUtils.findNearby(token, range, 'ally', {includeToken: true});
+    if (!near?.length || (near.length === 1 && near[0].id === token.id)) return;
+    let userId = socketUtils.firstOwner(actor, true);
+    let maxAmount = Math.max(1 + actor.system.abilities.int.mod, 2);
+    let minAmount = 2;
+    let targets = await dialogUtils.selectTargetDialog(
+        item.name, 
+        genericUtils.format('CHRISPREMADES.Macros.AdventurersAtlas.Use', {itemName: item.name, maxAmount, minAmount}), 
+        near, 
+        {type: 'multiple', maxAmount, minAmount, userId}
+    );
+    if (!targets || !targets[0]?.length) return;
+    await workflowUtils.syntheticItemRoll(item, targets[0]);
 }
 async function death({trigger: {actor}}) {
     let maps = effectUtils.getEffectByIdentifier(actor, 'adventurersAtlasCreator');
@@ -61,7 +58,7 @@ async function death({trigger: {actor}}) {
 }
 export let adventurersAtlas = {
     name: 'Adventurer\'s Atlas',
-    version: '1.5.16',
+    version: '1.5.17',
     rules: 'modern',
     midi: {
         item: [
@@ -105,6 +102,14 @@ export let adventurersAtlas = {
             label: 'CHRISPREMADES.Macros.AdventurersAtlas.RequireTools',
             type: 'checkbox',
             default: true,
+            category: 'homebrew',
+            homebrew: true
+        },
+        {
+            value: 'enforceRange',
+            label: 'CHRISPREMADES.Macros.AdventurersAtlas.Enforce5ft',
+            type: 'checkbox',
+            default: false,
             category: 'homebrew',
             homebrew: true
         }