+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 1.5.17 Release:
+## New Modern Content:
+- Ingenious Movement
+- Superior Atlas
+## Bug Fixes:
+- Generic Actions
+- Adventurer's Atlas
+- Guided Precision
+- Mapping Magic
+
 # 1.5.16 Release:
 ## New Legacy Content:
 - Armored Casing